refactor(service): moved each input adapter to its own module
